generos y test

+++ nefly/src/main/java/ar/com/ada/nefly/models/request/ContenidoRequest.java
@@ -8,7 +8,7 @@ public class ContenidoRequest {
     public String nombre;
     public String genero;
     public int año;
-    public int duracion;
+    public double duracion;
 
     public int nroEpisodio;
     public String titulo;
